Refactoring: split controller into multiple smaller controllers
The single `Amigo` controller was quite unwieldy and depended on
everything and the kitchen sink.
