+++ app/controllers/RoleController.scala
@@ -1,0 +1,14 @@
+package controllers
+
+import com.gu.googleauth.GoogleAuthConfig
+import data._
+import play.api.mvc._
+
+class RoleController(val authConfig: GoogleAuthConfig) extends Controller with AuthActions {
+
+  def listRoles = AuthAction {
+    Ok(views.html.roles(Roles.list))
+  }
+
+}
+
