+++ app/views/layout.scala.html
@@ -25,13 +25,13 @@
             <span class="icon-bar"></span>
             <span class="icon-bar"></span>
           </button>
-          <a class="navbar-brand" href="@routes.Amigo.index">AMIgo</a>
+          <a class="navbar-brand" href="@routes.RootController.index">AMIgo</a>
         </div>
         <div id="navbar" class="collapse navbar-collapse">
           <ul class="nav navbar-nav">
-            <li><a href="@routes.Amigo.baseImages">Base images</a></li>
-            <li><a href="@routes.Amigo.roles">Roles</a></li>
-            <li><a href="@routes.Amigo.recipes">Recipes</a></li>
+            <li><a href="@routes.BaseImageController.listBaseImages">Base images</a></li>
+            <li><a href="@routes.RoleController.listRoles">Roles</a></li>
+            <li><a href="@routes.RecipeController.listRecipes">Recipes</a></li>
           </ul>
         </div>
       </div>
