+++ app/views/newRecipe.scala.html
@@ -4,7 +4,7 @@
 
 @layout("AMIgo"){
 
-  @b3.form(routes.Amigo.createRecipe) {
+  @b3.form(routes.RecipeController.createRecipe) {
     @b3.text( form("id"), '_label -> "ID", 'placeholder -> "my-lovely-recipe" )
     @b3.textarea( form("description"), '_label -> "Description" )
     @b3.select( form("baseImageId"), availableBaseImages.map(_.id.value).zip(availableBaseImages.map(_.id.value)), '_label -> "Base image" )
