+++ app/views/showBaseImage.scala.html
@@ -6,7 +6,7 @@
   <div class="panel panel-default">
     <div class="panel-heading">Actions</div>
     <div class="panel-body">
-      <a class="btn btn-default" href="@routes.Amigo.editBaseImage(image.id)">Edit</a>
+      <a class="btn btn-default" href="@routes.BaseImageController.editBaseImage(image.id)">Edit</a>
     </div>
   </div>
